+++ CHANGELOG.md
@@ -22,6 +22,10 @@ Now, back to the program. Here are the changes for 1.0.0, with one especially no
   - Remove showing language and relevance data when auto detecting highlighted code block language
   - More minor bugfixes to permissions pages displays.
 
+### 0.11.4
+  - Bugfix for the message stating how many banners were deleted when deleting banners.
+  - Add an option to limit the total resolution of an image/video (width*height).
+
 ### 0.11.3
   - Fix max vs total upload count in controller for uploading board assets, flags and banners.
   - Move css theme assets to themes/assets instead of all lumped in one folder.